pyproject.toml
[tool.poetry]
name = "klavicle"
version = "0.1.0"
description = "Klaviyo Data Integration Tool"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "klavicle", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
rich = "^13.7.0"
klaviyo-api = "^5.3.1"
sqlalchemy = "^2.0.25"
pydantic = "^2.5.3"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
klavicle = "klavicle.__main__:cli"